added help function to lightcharge

+++ uboone/LLSelectionTool/OpT0Finder/Base/OpT0FinderTypes.h
@@ -161,7 +161,7 @@ namespace flashana {
     double y;
     double z;
     double q;
-    double plane;
+    int plane;
   };
 
   namespace msg {
